+++ src/Dialog/index.js
@@ -48,13 +48,6 @@ export default class Dialog extends Popup {
       onInit: () => {},
 
       /**
-       * Callback to run after component state is updated.
-       *
-       * @callback stateChangeCallback
-       */
-      onStateChange: () => {},
-
-      /**
        * Callback to run after the component is destroyed.
        *
        * @callback destroyCallback