+++ src/Disclosure/index.js
@@ -59,13 +59,6 @@ export default class Disclosure extends AriaComponent {
       onInit: () => {},
 
       /**
-       * Callback to run after component state is updated.
-       *
-       * @callback stateChangeCallback
-       */
-      onStateChange: () => {},
-
-      /**
        * Callback to run after the component is destroyed.
        *
        * @callback destroyCallback