+++ src/MenuBar/index.js
@@ -78,13 +78,6 @@ export default class MenuBar extends AriaComponent {
       onInit: () => {},
 
       /**
-       * Callback to run after component state is updated.
-       *
-       * @callback stateChangeCallback
-       */
-      onStateChange: () => {},
-
-      /**
        * Callback to run after the component is destroyed.
        *
        * @callback destroyCallback